PAALALA GUI
A software application that was created as a final project in one of courses.

Paalala is a software application that serves as a alarm for medications.

The version of PAALALA that was commited is only v.0.1, that is why there are some things that are not yet working and needs debugging.